feat: add hook
